requirement.txt
streamlit
PIL
opencv-python
requests
dotenv
google-generativeai
langchain
pandas

